fix and move spline
